# docs/api/driver-app.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RouteShareApp Driver API",
    "version": "0.1.0",
    "description": "OpenAPI contract consumed by the Driver mobile app. Authentication is handled by Keycloak using OIDC Authorization Code + PKCE; this API validates JWT bearer tokens."
  },
  "servers": [
    {
      "url": "https://api.routeshare.local",
      "description": "Local/dev API gateway"
    }
  ],
...
  "paths": {
...
    "/api/v1/driver/application": {
      "post": {
        "tags": [
          "Driver Profile"
        ],
        "summary": "Apply as driver",
        "operationId": "DriverProfileApplyAsDriver",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverProfile"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverApplicationRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/driver/profile": {
      "get": {
        "tags": [
          "Driver Profile"
        ],
        "summary": "Get driver profile",
        "operationId": "DriverProfileGetDriverProfile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverProfile"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/driver/vehicles": {
      "post": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Create vehicle",
        "operationId": "VehiclesCreateVehicle",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVehicleRequest"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Vehicles"
        ],
        "summary": "List vehicles",
        "operationId": "VehiclesListVehicles",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Vehicle"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Keycloak JWT access token."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message",
          "correlationId"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "DriverProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "fullName": {
            "type": "string"
          },
          "kycStatus": {
            "type": "string"
          },
          "driverStatus": {
            "type": "string"
          },
          "approvalStatus": {
            "type": "string"
          },
          "ratingAverage": {
            "type": "number"
          },
          "totalCompletedTrips": {
            "type": "integer"
          }
        }
      },
      "DriverApplicationRequest": {
        "type": "object",
        "required": [
          "fullName",
          "dateOfBirth",
          "addressText"
        ],
        "properties": {
          "fullName": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "gender": {
            "type": "string"
          },
          "addressText": {
            "type": "string"
          }
        }
      },
...
      "Vehicle": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "make": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "year": {
            "type": "integer"
          },
          "color": {
            "type": "string"
          },
          "registrationNumber": {
            "type": "string"
          },
          "passengerSeatCapacity": {
            "type": "integer"
          },
          "features": {
            "type": "object",
            "additionalProperties": true
          },
          "verificationStatus": {
            "type": "string"
          },
          "activeStatus": {
            "type": "string"
          },
          "isPrimary": {
            "type": "boolean"
          }
        }
      },
      "CreateVehicleRequest": {
        "type": "object",
        "required": [
          "make",
          "model",
          "year",
          "color",
          "registrationNumber",
          "passengerSeatCapacity"
        ],
        "properties": {
          "make": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "year": {
            "type": "integer"
          },
          "color": {
            "type": "string"
          },
          "registrationNumber": {
            "type": "string"
          },
          "vehicleType": {
            "type": "string"
          },
          "passengerSeatCapacity": {
            "type": "integer"
          },
          "features": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}